Added support for describe generics params in nested gererics in Shorthand format of type for Configuration.
For example Func[List[int], string] must be resolved as Func<List<int>, string> (before it's resolved as Func<List`1, string>.

The problem was in TypeResolverImpl, when we are resolve genericParams, we are ignore internal genericParams (because FullName don't get them).
I add new property FullNameWithNestedGenerics for this case.

Also added some tests for this.